moved get_component and get_component_system benches into world_get.rs into query_get_component_simple
