+++ benches/benches/bevy_ecs/components/mod.rs
@@ -5,8 +5,6 @@ mod add_remove_big_table;
 mod add_remove_sparse_set;
 mod add_remove_table;
 mod archetype_updates;
-mod get_component;
-mod get_component_system;
 mod insert_simple;
 mod insert_simple_unbatched;
 
@@ -16,7 +14,6 @@ criterion_group!(
     components_benches,
     add_remove,
     add_remove_big,
-    get_component,
     insert_simple,
     no_archetypes,
     added_archetypes,
@@ -47,21 +44,6 @@ fn add_remove_big(c: &mut Criterion) {
     });
     group.bench_function("sparse_set", |b| {
         let mut bench = add_remove_big_sparse_set::Benchmark::new();
-        b.iter(move || bench.run());
-    });
-    group.finish();
-}
-
-fn get_component(c: &mut Criterion) {
-    let mut group = c.benchmark_group("get_component");
-    group.warm_up_time(std::time::Duration::from_millis(500));
-    group.measurement_time(std::time::Duration::from_secs(4));
-    group.bench_function("base", |b| {
-        let mut bench = get_component::Benchmark::new();
-        b.iter(move || bench.run());
-    });
-    group.bench_function("system", |b| {
-        let mut bench = get_component_system::Benchmark::new();
         b.iter(move || bench.run());
     });
     group.finish();
